feature(mailbox-generator): move files
