+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/core/domain/parsing/email/EmailParsingException.java
@@ -1,4 +1,4 @@
-package de.dhbw.karlsruhe.students.mailflow.core.application.email.parsing;
+package de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.email;
 
 public class EmailParsingException extends RuntimeException {
     public EmailParsingException(String errorMessage, Throwable err) {
