+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/core/domain/parsing/mailbox/MailboxRepository.java
@@ -1,10 +1,10 @@
-package de.dhbw.karlsruhe.students.mailflow.core.application.email;
+package de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox;
 
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.Mailbox;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.enums.MailboxType;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.value_objects.Address;
-import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.FileCreationException;
-import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.FileWritingException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator.FileCreationException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator.FileWritingException;
 import java.io.File;
 import java.util.Optional;
