+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/core/domain/parsing/mailbox/generator/MailboxGenerator.java
@@ -1,4 +1,4 @@
-package de.dhbw.karlsruhe.students.mailflow.core.domain.parsing;
+package de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator;
 
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.Mailbox;
