+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/core/domain/parsing/mailbox/reader/MailboxParser.java
@@ -1,4 +1,4 @@
-package de.dhbw.karlsruhe.students.mailflow.core.application.email.parsing;
+package de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.reader;
 
 
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.Mailbox;
