+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/core/domain/parsing/mailbox/reader/MailboxParsingException.java
@@ -1,4 +1,4 @@
-package de.dhbw.karlsruhe.students.mailflow.core.application.email.parsing;
+package de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.reader;
 
 public class MailboxParsingException extends RuntimeException {
   public MailboxParsingException(String errorMessage, Throwable err) {
