+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/external/infrastructure/email/EmailMetadataFactory.java
@@ -1,6 +1,6 @@
 package de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email;
 
-import de.dhbw.karlsruhe.students.mailflow.core.application.email.parsing.EmailParsingException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.email.EmailParsingException;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.value_objects.*;
 import jakarta.mail.Message;
 import jakarta.mail.Message.RecipientType;
