+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/external/infrastructure/email/JSONMailboxGenerator.java
@@ -3,7 +3,7 @@ package de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email;
 import static de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email.GSONConfiguration.getConfiguredGson;
 
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.Mailbox;
-import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.MailboxGenerator;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator.MailboxGenerator;
 
 public class JSONMailboxGenerator implements MailboxGenerator {
