+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/external/infrastructure/email/JSONMailboxParser.java
@@ -2,8 +2,8 @@ package de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email;
 
 import static de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email.GSONConfiguration.getConfiguredGson;
 
-import de.dhbw.karlsruhe.students.mailflow.core.application.email.parsing.MailboxParser;
-import de.dhbw.karlsruhe.students.mailflow.core.application.email.parsing.MailboxParsingException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.reader.MailboxParser;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.reader.MailboxParsingException;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.Mailbox;
 
 public class JSONMailboxParser implements MailboxParser {
