+++ src/test/java/de/dhbw/karlsruhe/students/mailflow/unit/LocalMailboxRepositoryTest.java
@@ -1,11 +1,11 @@
 package de.dhbw.karlsruhe.students.mailflow.unit;
 
-import de.dhbw.karlsruhe.students.mailflow.core.application.email.MailboxRepository;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.MailboxRepository;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.Mailbox;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.enums.MailboxType;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.value_objects.Address;
-import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.FileCreationException;
-import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.FileWritingException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator.FileCreationException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.parsing.mailbox.generator.FileWritingException;
 import de.dhbw.karlsruhe.students.mailflow.external.infrastructure.email.LocalMailboxRepository;
 import java.io.File;
 import java.util.List;
